// src/app/planeacion/interfaces/Planecion.ts
export interface ITproducto {
  nomProveedor:string; 
  rfc: string;
  noIdentificacion: string;
  codarticulo: string;
  umedida: string;
  uds: string;
  descripcion: string;
  refproveedor: string;
  medidareferencia: string;
  unidadmedida: string;
  planeacion: boolean; 
  p_uds:string; 
  p_umedida:string; 
  i_uds:string; 
  i_umedida:string; 
}